+++ src/cli/clipboard.rs
@@ -2,25 +2,20 @@
     unix,
     not(any(target_os = "macos", target_os = "android", target_os = "emscripten"))
 ))]
-fn is_program_in_path(program: &str) -> bool {
-    use std::{env, fs};
-
-    if let Ok(path) = env::var("PATH") {
-        for p in path.split(':') {
-            let p_str = format!("{}/{}", p, program);
-            if fs::metadata(p_str).is_ok() {
-                return true;
+// use xclip (or a similar program that is installed) because the kernel deletes the clipboard after the process ends
+pub fn set_clipboard(content: String) {
+    fn is_program_in_path(program: &str) -> bool {
+        if let Ok(path) = std::env::var("PATH") {
+            for p in path.split(':') {
+                let p_str = format!("{}/{}", p, program);
+                if std::fs::metadata(p_str).is_ok() {
+                    return true;
+                }
             }
         }
+        false
     }
-    false
-}
 
-#[cfg(all(
-    unix,
-    not(any(target_os = "macos", target_os = "android", target_os = "emscripten"))
-))]
-pub fn set_clipboard(content: String) {
     use std::{
         io::Write,
         process::{Command, Stdio},
@@ -38,6 +33,7 @@ pub fn set_clipboard(content: String) {
             .stdin(Stdio::piped())
             .spawn()
             .expect("execute command xsel")
+
     // xclip
     } else if is_program_in_path("xclip") {
         child = Command::new("xclip")
@@ -47,41 +43,43 @@ pub fn set_clipboard(content: String) {
             .stdin(Stdio::piped())
             .spawn()
             .expect("execute command xclip")
+
     // termux
     } else if is_program_in_path("termux-clipboard-set") {
         child = Command::new("termux-clipboard-set")
             .stdin(Stdio::piped())
             .spawn()
             .expect("execute command termux-clipboard-set")
+
+    // the above programs responsible for the clipboard were not found
     } else {
         println!(
             "{} {}",
             "WARN".yellow(),
             "command for clipboard not found".magenta()
         );
-        return;
+
+        return
     }
 
+    // copy the content (send it to stdin command)
     child
         .stdin
         .as_mut()
         .unwrap()
         .write_all(content.as_bytes())
         .expect("execute command");
-    child.wait_with_output().unwrap();
+
+    child
+        .wait_with_output()
+        .expect("wait for clipboard command output");
 }
 
 #[cfg(not(all(
     unix,
     not(any(target_os = "macos", target_os = "android", target_os = "emscripten"))
 )))]
-use arboard::Clipboard;
-
-#[cfg(not(all(
-    unix,
-    not(any(target_os = "macos", target_os = "android", target_os = "emscripten"))
-)))]
 pub fn set_clipboard(content: String) {
-    let mut clipboard = Clipboard::new().unwrap();
-    clipboard.set_text(content).unwrap();
+    let mut clipboard = arboard::Clipboard::new().unwrap();
+    clipboard.set_text(content).execute(format!("set clipboard to '{content}'"));
 }