+++ src/cli/delete_image.rs
@@ -1,11 +1,15 @@
 use colored::Colorize;
-
-use imgurs::api::{delete_image::delete_image as del_img, ImgurClient};
+use imgurs::ImgurClient;
 
 pub async fn delete_image(client: ImgurClient, delete_hash: String) {
-    let i = del_img(client, delete_hash)
+    // delete image from imgur
+    client
+        .delete_image(delete_hash)
         .await
         .expect("send api request");
 
-    println!("{}", i.magenta());
+    println!(
+        "{}",
+        "If Delete Hash was correct the image was deleted!".magenta()
+    );
 }