+++ src/cli/parse.rs
@@ -1,10 +1,11 @@
 use clap::{Command, IntoApp, Parser, Subcommand};
 use clap_complete::{generate, Generator, Shell};
-use imgurs::api::ImgurClient;
+use imgurs::ImgurClient;
 use std::io::{self, stdout};
 
 use crate::cli::{credits::*, delete_image::*, info_image::*, upload_image::*};
 
+// get program name and varsion from Cargo.toml
 const VERSION: Option<&str> = option_env!("CARGO_PKG_VERSION");
 const NAME: Option<&str> = option_env!("CARGO_PKG_NAME");
 
@@ -21,7 +22,7 @@ struct Cli {
 
 #[derive(Subcommand, Debug)]
 enum Commands {
-    #[clap(about = "Print API Rate Limit", display_order = 1)]
+    #[clap(about = "Print Client Rate Limit", display_order = 1)]
     Credits,
 
     #[clap(about = "Upload image to Imgur", display_order = 2)]
@@ -37,7 +38,7 @@ enum Commands {
         about = "Generate completion file for a shell [bash, elvish, fish, powershell, zsh]",
         display_order = 5
     )]
-    Completions { shell: String },
+    Completions { shell: Shell },
 
     #[clap(about = "Generate man page", display_order = 6)]
     Manpage,
@@ -50,41 +51,26 @@ fn print_completions<G: Generator>(gen: G, app: &mut Command) {
 pub async fn parse(client: ImgurClient) {
     let args = Cli::parse();
 
-    match &args.command {
-        Commands::Credits => {
-            credits(client).await;
-        }
+    match args.command {
+        Commands::Credits => credits(client).await,
 
-        Commands::Upload { path } => {
-            upload_image(client, path).await;
-        }
+        Commands::Upload { path } => upload_image(client, path.to_string()).await,
 
-        Commands::Delete { delete_hash } => {
-            delete_image(client, delete_hash.to_string()).await;
-        }
+        Commands::Delete { delete_hash } => delete_image(client, delete_hash.to_string()).await,
 
-        Commands::Info { id } => {
-            image_info(client, id).await;
-        }
+        Commands::Info { id } => image_info(client, id.to_string()).await,
 
         Commands::Completions { shell } => {
             let mut app = Cli::command();
 
-            match shell.as_str() {
-                "bash" => print_completions(Shell::Bash, &mut app),
-                "elvish" => print_completions(Shell::Elvish, &mut app),
-                "fish" => print_completions(Shell::Fish, &mut app),
-                "powershell" => print_completions(Shell::PowerShell, &mut app),
-                "zsh" => print_completions(Shell::Zsh, &mut app),
-
-                _ => panic!("Completions to shell `{shell}`, not found!"),
-            }
+            print_completions(shell, &mut app)
         }
 
         Commands::Manpage => {
             let clap_app = Cli::command();
             let man = clap_mangen::Man::new(clap_app);
-            man.render(&mut io::stdout()).unwrap();
+
+            man.render(&mut io::stdout()).expect("generate manpage")
         }
     }
 }