+++ src/cli/upload_image.rs
@@ -1,14 +1,12 @@
 use super::clipboard::set_clipboard;
-use imgurs::api::{upload_image::upload_image as upload_img, ImgurClient};
+use imgurs::ImgurClient;
 use notify_rust::Notification;
 
 use crate::{cli::webhook::send_discord_webhook, config::toml};
 
 use super::print_image_info;
 
-use base64::encode as base64_encode;
-use std::{fs::read as fs_read, path::Path};
-
+// show notification
 macro_rules! notify (
     ($notification: expr) => (
         if toml::parse().notification.enabled {
@@ -17,43 +15,39 @@ macro_rules! notify (
     );
 );
 
-pub async fn upload_image(client: ImgurClient, path: &str) {
-    let mut image: String = path.to_string();
-
+pub async fn upload_image(client: ImgurClient, path: String) {
+    // parse configuration file
     let config = toml::parse();
 
-    if Path::new(path).exists() {
-        let bytes = fs_read(path)
-            .map_err(|err| err.to_string())
-            .expect("read file");
-        image = base64_encode(bytes);
-    } else if !validator::validate_url(path) {
-        panic!("{path} is not a url")
-    }
-
-    let mut i = upload_img(client, &image).await.unwrap_or_else(|err| {
+    // upload a image to imgur
+    let mut i = client.upload_image(path).await.unwrap_or_else(|err| {
         notify!(Notification::new()
             .summary("Error!")
             .body(&format!("Error: {}", &err.to_string()))
             .appname("Imgurs")); // I don't think you can set it to error
 
-        panic!("{}", err)
+        panic!("send request to imagur api: {}", err)
     });
 
+    // change domain to proxy (to be set in config)
     if config.imgur.image_cdn != "i.imgur.com" {
-        i.data.link = i.data.link.replace("i.imgur.com", "cdn.magicuser.cf")
+        i.data.link = i.data.link.replace("i.imgur.com", &config.imgur.image_cdn)
     }
 
+    // print image information from imgur
     print_image_info(i.clone());
 
-    let body = format!("Uploaded {}", i.data.link);
+    // send notification that the image has been uploaded
+    notify!(Notification::new()
+        .summary("Imgurs")
+        .body(&format!("Uploaded {}", i.data.link)));
 
-    notify!(Notification::new().summary("Imgurs").body(&body));
-
+    // if enabled copy link to clipboard
     if config.clipboard.enabled {
         set_clipboard(i.data.link.clone())
     }
 
+    // if enabled send embed with link and deletehash to discord (something like logger)
     if config.discord_webhook.enabled {
         send_discord_webhook(i.data.link, i.data.deletehash.unwrap())
             .await