+++ src/cli/webhook.rs
@@ -3,13 +3,21 @@ use std::error::Error;
 
 use crate::config::toml;
 
+// send embed with link and deletehash to discord (something like logger)
 pub async fn send_discord_webhook(
     link: String,
     deletehash: String,
-) -> Result<bool, Box<dyn Error + Send + Sync>> {
+) -> Result<(), Box<dyn Error + Send + Sync>> {
+    // get discord webhook uri from config
     let url = toml::parse().discord_webhook.uri;
+
+    // create WebhookClient
     let client: WebhookClient = WebhookClient::new(&url);
 
+    // get program version
+    let version = option_env!("CARGO_PKG_VERSION").unwrap_or("unknown");
+
+    // send discord webhook
     client
         .send(|message| {
             message.username("Imgurs").embed(|embed| {
@@ -17,14 +25,10 @@ pub async fn send_discord_webhook(
                     .title(&link)
                     .description(&format!("Delete Hash ||{deletehash}||"))
                     .image(&link)
-                    .footer(
-                        &format!(
-                            "Imgurs v{}",
-                            option_env!("CARGO_PKG_VERSION").unwrap_or("unknown")
-                        ),
-                        None,
-                    )
+                    .footer(&format!("Imgurs v{version}"), None)
             })
         })
-        .await
+        .await?;
+
+    Ok(())
 }