+++ src/lib.rs
@@ -1,1 +1,3 @@
-pub mod api;
+mod api;
+
+pub use api::*;